pyproject.toml
[tool.poetry]
name = "datalake2"
version = "0.1.0"
description = ""
authors = ["David Tacite <david.tacite@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
azure-identity = "^1.19.0"
azure-keyvault = "^4.2.0"
azure-storage-blob = "^12.24.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
